contract size too large

+++ contracts/PoolsharkHedgePoolFactory.sol
@@ -64,7 +64,7 @@ contract PoolsharkHedgePoolFactory is
                     token0,
                     token1,
                     uint24(swapFee),
-                    uint24(tickSpacing)
+                    tickSpacing
                 )
             )
         );
@@ -73,7 +73,7 @@ contract PoolsharkHedgePoolFactory is
         poolList.push(pool);
 
         // emit event for indexers
-        emit PoolCreated(token0, token1, uint24(swapFee), uint24(tickSpacing), pool);
+        emit PoolCreated(token0, token1, uint24(swapFee), tickSpacing, pool);
     }
 
     function getHedgePool(
